+++ tests/utils.py
@@ -97,7 +97,7 @@ class SparkSessionMock:
 
         def get(self, key):
             if key not in self._conf:
-                raise Exception("NoSuchElementException")
+                raise NameError(key)
             return self._conf[key]
 
         def set(self, key, value):