+++ CHANGES.md
@@ -2,7 +2,8 @@
 _All notable changes to the codebase are documented in this file._
 
 ## [0.6.5] - 08.10.2023
-- Fixed: `NoSuchElementException` java exception may belong not only to `java.util` namespace.
+- Fixed: some functions of `dtflw.databricks` used to catch and filter for `java.util.NoSuchElementException`.
+Filtering has been removed since it may be a different exception class.
 
 ## [0.6.4] - 09.04.2023
 - Improve validation, skip reading output table if table expected_columns are not set.